Moved the PSNR-B from no_ref to full_ref
