+++ sewar/no_ref.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, division, print_function
 import numpy as np
-from math import log2, log10
 from .full_ref import uqi
 from scipy.ndimage.filters import uniform_filter
 from scipy.misc import imresize
@@ -77,102 +76,3 @@ def qnr (pan,ms,fused,alpha=1,beta=1,p=1,q=1,r=4,ws=7):
 	return a*b
 
 
-def compute_bef(im, block_size=8):
-	"""Calculates Blocking Effect Factor (BEF) for a given grayscale/one channel image
-
-	C. Yim and A. C. Bovik, "Quality Assessment of Deblocked Images," in IEEE Transactions on Image Processing,
-		vol. 20, no. 1, pp. 88-98, Jan. 2011.
-
-	:param im: input image (numpy ndarray)
-	:param block_size: Size of the block over which DCT was performed during compression
-	:return: float -- bef.
-	"""
-	if len(im.shape) == 3:
-		height, width, channels = im.shape
-	elif len(im.shape) == 2:
-		height, width = im.shape
-		channels = 1
-	else:
-		raise ValueError("Not a 1-channel/3-channel grayscale image")
-
-	if channels > 1:
-		raise ValueError("Not for color images")
-
-	h = np.array(range(0, width - 1))
-	h_b = np.array(range(block_size - 1, width - 1, block_size))
-	h_bc = np.array(list(set(h).symmetric_difference(h_b)))
-
-	v = np.array(range(0, height - 1))
-	v_b = np.array(range(block_size - 1, height - 1, block_size))
-	v_bc = np.array(list(set(v).symmetric_difference(v_b)))
-
-	d_b = 0
-	d_bc = 0
-
-	# h_b for loop
-	for i in list(h_b):
-		diff = im[:, i] - im[:, i+1]
-		d_b += np.sum(np.square(diff))
-
-	# h_bc for loop
-	for i in list(h_bc):
-		diff = im[:, i] - im[:, i+1]
-		d_bc += np.sum(np.square(diff))
-
-	# v_b for loop
-	for j in list(v_b):
-		diff = im[j, :] - im[j+1, :]
-		d_b += np.sum(np.square(diff))
-
-	# V_bc for loop
-	for j in list(v_bc):
-		diff = im[j, :] - im[j+1, :]
-		d_bc += np.sum(np.square(diff))
-
-	# N code
-	n_hb = height * (width/block_size) - 1
-	n_hbc = (height * (width - 1)) - n_hb
-	n_vb = width * (height/block_size) - 1
-	n_vbc = (width * (height - 1)) - n_vb
-
-	# D code
-	d_b /= (n_hb + n_vb)
-	d_bc /= (n_hbc + n_vbc)
-
-	# Log
-	if d_b > d_bc:
-		t = log2(block_size)/log2(min(height, width))
-	else:
-		t = 0
-
-	# BEF
-	bef = t*(d_b - d_bc)
-
-	return bef
-
-
-def psnrb(GT, P):
-	"""Calculates PSNR with Blocking Effect Factor for a given pair of images (PSNR-B)
-
-	:param GT: first (original) input image in YCbCr format or Grayscale.
-	:param P: second (corrected) input image in YCbCr format or Grayscale..
-	:return: float -- psnr_b.
-	"""
-	if len(GT.shape) == 3:
-		GT = GT[:, :, 0]
-
-	if len(P.shape) == 3:
-		P = P[:, :, 0]
-
-	imdff = np.double(GT) - np.double(P)
-
-	mse = np.mean(np.square(imdff.flatten()))
-	bef = compute_bef(P)
-	mse_b = mse + bef
-
-	if np.amax(P) > 2:
-		psnr_b = 10 * log10(255**2/mse_b)
-	else:
-		psnr_b = 10 * log10(1/mse_b)
-
-	return psnr_b
